hcm: split out filter management into seperate class
